test: setup jest
